Use Laravel Gates to authorize access to user

+++ app/Http/Controllers/QuestionsController.php
@@ -72,7 +72,13 @@ class QuestionsController extends Controller
      */
     public function edit(Question $question)
     {
+
+        if (\Gate::denies('update-question', $question)) {
+            abort(403, "Access Denied");
+        }
+        
         return view("questions.edit")->with('question', $question);
+
     }
 
     /**
@@ -84,6 +90,10 @@ class QuestionsController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+        if (\Gate::denies('update-question', $question)) {
+            abort(403, "Access Denied");
+        }
+        
         $question->update($request->only('title', 'body'));
 
         return redirect()->route('questions.index')->with('success', "Your question is updated");
@@ -97,6 +107,11 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
+
+        if (\Gate::denies('delete-question', $question)) {
+            abort(403, "Access Denied");
+        }
+
         $question->delete();
 
         return redirect()->route('questions.index')->with('success', "Your question is deleted");
